Consistently use .mjs to denote modules
The use of `.mjs` to denote modules as oppoosed to classic scripts is considered to be a best practice.

In practice, Node.js requires `.mjs` for modules, and on the web file extensions doesn't matter. On top of any host environment requirements, there is value in clearly making the distinction between modules and classic scripts during development. Dedicated file extensions for each type of script accomplish this. Thus, the most pragmatic solution nowadays is to use `.mjs` for modules consistently regardless of the platform.

https://developers.google.com/web/fundamentals/primers/modules#mjs