Start a new approach, based on searchable invisible DOM
This branch contains a new take on the problem space, based around the concepts of (and current flagged implementation of) searchable invisible DOM [1]. Searchable invisible DOM allows virtualized content to be accessible, work with in-page anchors, find in page, etc.

Given this foundational technology, this branch starts over on the API design side of the virtual scroller. The foundational piece, found in this commit, is a new <virtual-content> element. This element has no real API: it just manages its children's searchable-invisibleness. Thus we don't have any of the master branch's <virtual-scroller> API for ItemSources, creating and recycling elements, etc. See #147 for more info.

Closes #147.

[1]: https://github.com/rakina/searchable-invisible-dom

Co-authored-by: Russell Bicknell <bicknellr@google.com>
Co-authored-by: Domenic Denicola <d@domenic.me>