Find the nearest scrolling ancestor by walking the tree rather than using scroll event targets. (...)
The target of scroll events isn't always the element that needs to have its
`scrollTop` corrected when height estimates are found to be incorrect during an
update. This can happen if the virtual content is nested in multiple scrollable
elements and the user scrolls more of the virtual content into view using one
of the outer scrollable elements.