Handle resizes of the nearest scroll container by watching for viewport intersection with unrendered space. (...)
Before this change, the virtual-content did not update if its nearest scroll
container (possibly the window) resized in a way that made more of the
virtual-content visible. To fix this, sentinel elements sized to fill the space
between visible elements are inserted and monitored with an
IntersectionObserver, using the viewport as the root. This has the convenient
side effect of covering all of the cases that were previously handled by
listening to scroll events. Additionally, this prevents any virtual-content
elements that do not intersect the viewport from updating for scroll events.