Make README an explainer
This commit is the first step at cleaning up this repository to focus on purely being a proposal for a new web platform feature. Until now, it was a confusing mix of web platform feature explainer, and Chromium-specific prototype code. This focuses the repository purely on being a web platform feature repository, removing all references to the prototype work, which will instead move to the Chromium source tree.

See https://github.com/valdrinkoshi/virtual-scroller/issues/158#issuecomment-480913127 for a bit more explanation and background.